requirements.txt
# Exoplanet Detection with AI/ML - Requirements
# Core data science packages
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Machine learning packages
scikit-learn>=1.1.0
xgboost>=1.6.0
catboost>=1.1.0

# Additional utilities
joblib>=1.2.0

# Web application
flask>=2.3.0
flask-cors>=4.0.0

# Environment variables
python-dotenv>=1.0.0

# LLM for habitability analysis
openai>=1.0.0

# Database
psycopg2-binary>=2.9.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: for better performance
numba>=0.56.0
